+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.7.20
+
+* less CPU usage in the stdout rewriter thread